refactor: extract TaskMobileCard and FilterControls components
- Extract TaskMobileCard to separate file with React.memo for performance
- Create shared FilterControls component to reduce code duplication
- Add accessibility attributes (role, aria-label) to mobile filter scroller
- Add clarifying comment for hasActiveFilters logic

Addresses code review feedback on PR #379.
